# Memory corruption fix