Fixes #2159