@YohDeadfall 2018-09-22
```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17754
Intel Xeon CPU E5-2687W 0 3.10GHz, 2 CPU, 32 logical and 16 physical cores
.NET Core SDK=2.1.402
  [Host]     : .NET Core 2.1.4 (CoreCLR 4.6.26814.03, CoreFX 4.6.26814.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.4 (CoreCLR 4.6.26814.03, CoreFX 4.6.26814.02), 64bit RyuJIT
```

## Int16

### After

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|--------|------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 34.78 ns | 0.7643 ns | 0.6382 ns | 28,752,616.4 |      - |       0 B |
|  Write |     0 | 89.43 ns | 1.7158 ns | 1.4327 ns | 11,181,791.1 | 0.0088 |      56 B |

### Before

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|--------|------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 32.40 ns | 0.7477 ns | 1.4043 ns | 30,862,603.4 |      - |       0 B |
|  Write |     0 | 86.00 ns | 0.8631 ns | 0.8073 ns | 11,628,414.5 | 0.0088 |      56 B |

## Int32

### After

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------- |------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 34.08 ns | 0.4486 ns | 0.3503 ns | 29,339,539.1 |      - |       0 B |
|  Write |     0 | 89.08 ns | 1.3825 ns | 1.1544 ns | 11,225,927.4 | 0.0100 |      64 B |

### Before

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------- |------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 32.49 ns | 0.7487 ns | 1.0970 ns | 30,777,529.7 |      - |       0 B |
|  Write |     0 | 89.71 ns | 1.8428 ns | 2.1220 ns | 11,146,702.5 | 0.0100 |      64 B |

## Int64

### After

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------- |------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 38.75 ns | 0.8280 ns | 0.9857 ns | 25,806,620.0 |      - |       0 B |
|  Write |     0 | 88.00 ns | 1.8606 ns | 2.6684 ns | 11,363,785.6 | 0.0100 |      64 B |

### Before

| Method | Value |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------- |------ |---------:|----------:|----------:|-------------:|-------:|----------:|
|   Read |     0 | 38.14 ns | 0.7562 ns | 0.7073 ns | 26,216,932.3 |      - |       0 B |
|  Write |     0 | 97.75 ns | 1.8530 ns | 1.6427 ns | 10,230,281.4 | 0.0100 |      64 B |