@austindrenski 2018-09-22
@YohDeadfall I haven't worked with [this feature](http://mattwarren.org/2016/03/09/adventures-in-benchmarking-method-inlining/), but BenchmarkDotNet should provide some output on whether inlining is happening or not. 

Could you take a look on whether the bounds check has any impact on the inlining?