@YohDeadfall 2018-09-23
@austindrenski Good idea! Will do it and getting disassembly too.